net/filter-mirror: add enable_when_stopped option for filter-redirector
Add a QAPI option 'enable_when_stopped' for filter-redirector that
allows it to continue receiving data from netdev when VM is stopped.

When enabled, a VM state change handler activates the netdev's
read_poll() when VM stops, allowing the netdev to continue reading
data and passing it through filter-redirector to the chardev output.

This is useful for scenarios where filter-redirector needs to
process network traffic even when the VM is paused.

Usage:
  -object filter-redirector,id=f0,netdev=net0,queue=tx,\
          outdev=chardev0,enable_when_stopped=true

Signed-off-by: Jason Wang <jasowang@redhat.com>