netfilter: enable read_poll when redirector created while stopped
filter-redirector uses a vm change state handler to enable netdev read_poll()
when enable_when_stopped=true and the VM stops, so that host traffic can be
still drained into the netfilter chain while vCPUs are not running.

However, qemu_add_vm_change_state_handler() does not invoke the callback for
the current VM state at registration time. If a redirector is created after
the VM is already stopped (or if enable_when_stopped is toggled on while
stopped), read_poll() may remain disabled and packets will not be drained.

Add a helper that checks the current runstate and proactively enables
read_poll() when enable_when_stopped is set and the VM is already stopped.
Call it from redirector setup and from the enable_when_stopped property setter.

Signed-off-by: Jason Wang <jasowang@redhat.com>